Hex/images/requirements.txt
django>=4.0.0,<4.1.0
djangorestframework
requests
easy-thumbnails
django-cleanup
django-request-token///////////
opencv-python